requirements.txt
# Existing ML dependencies
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
openpyxl>=3.1.0

# New LLM and CLI dependencies
openai>=1.12.0
google-generativeai>=0.3.0
anthropic>=0.18.0
python-dotenv>=1.0.0
rich>=13.7.0
sqlalchemy>=2.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
requests>=2.31.0
